Overhaul dashboard
Use search instead of listing all jobs in the database.
Display parameters for each job in the info panel.
Allow diffing source code of different jobs.

Tagging jobs, while in the UI, is not yet supported by the
database. In addition, abstractions on top of the database are
introduced that should be reused in the job-launching portion of
nbjob.